app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
anthropic==0.42.0
pydantic[email]==2.10.4
bcrypt==4.2.1
PyJWT==2.10.1
python-dotenv==1.0.1
